added transcription for youtube loom and awssome

+++ app/services/awesomess_transcript.py
@@ -1,7 +1,8 @@
 import requests
 from urllib.parse import urlparse, parse_qs
-
-# https://www.awesomescreenshot.com/api/v1/video/load_video?id=29984315&key=56e0915033f19b126f9903de6c0dd881
+import os
+import whisper
+from app.utils.utils import has_audio_stream
 
 def get_awesomess_transcript(url: str) -> str:
     try:
@@ -23,23 +24,35 @@ def get_awesomess_transcript(url: str) -> str:
         video_url = json_data.get("data", {}).get("video", {}).get("fileWebMURI")
 
         if not video_url:
-            raise ValueError("webmOrigURI not found in response")
+            raise ValueError("fileWebMURI not found in response")
 
-        output_file = "video_downloaded.mp4"
+        output_file = f"video/video_{video_id}.webm"
 
         headers = {
             'User-Agent': 'Mozilla/5.0', 
         }
 
-        response = requests.get(video_url, headers=headers, stream=True)
+        if not os.path.exists(output_file):
+            os.makedirs(os.path.dirname(output_file), exist_ok=True)
+            with requests.get(video_url, headers=headers, stream=True) as r:
+                r.raise_for_status()
+                with open(output_file, 'wb') as f:
+                    for chunk in r.iter_content(chunk_size=8192):
+                        if chunk:
+                            f.write(chunk)
 
-        with open(output_file, 'wb') as f:
-            for chunk in response.iter_content(chunk_size=8192):
-                if chunk:
-                    f.write(chunk)
+        if not has_audio_stream(output_file):
+            os.remove(output_file)
+            return "No audio stream found in Loom video — transcription not possible."                    
 
-        print("Video downloaded successfully.")
-        return video_url
+        model = whisper.load_model("base", device="cpu")
+        result = model.transcribe(output_file,fp16=False)
+        transcript = result.get("text", "")
+
+        if os.path.exists(output_file):
+            os.remove(output_file)
+
+        return transcript
 
     except requests.RequestException as req_err:
         print(f"[ERROR] Request failed: {req_err}")
